Fix MPS issues
Okay, this one is super weird. MPS produces NaN values
from the embedding layer, loss, and everything in between,
__unless__ there is a no-op which forces evaluation
after the attention layer - e.g. `x.clone()`, `x + 0`
or even `print(x)`.

This is NOT a problem with CPU. I can NOT explain it.
It just is (it is, is, is).